armdevtools update for docker
added dev tools rebuild for the arm docker image
