+++ devtools/armui.py
@@ -9,10 +9,12 @@ import os
 import log
 
 
-# Stop the ARM UI
-#  INPUT: none
-#  OUTPUT: none
 def stop():
+    """
+    Stop the ARM UI
+        INPUT: none
+        OUTPUT: none
+    """
     try:
         log.info("Going to stop ARMUI - requesting sudo")
         os.system("sudo systemctl stop armui.service")
@@ -25,9 +27,36 @@ def stop():
 #  INPUT: none
 #  OUTPUT: none
 def start():
+    """
+    Stop the ARM UI
+        INPUT: none
+        OUTPUT: none
+    """
     try:
         log.info("Going to restart ARMUI - requesting sudo")
         os.system("sudo systemctl start armui.service")
         log.success("ARM UI started")
     except Exception as error:
         log.error(f"ARM UI unable to start - {error}")
+
+
+def run_command(command, statement):
+    """
+    Run os commands and check they run
+        INPUT: STRING command, STRING statment
+        OUTPUT: none
+    """
+    try:
+        # Stop ARM container
+        log.info("-------------------------------------")
+        log.info(f"Executing: {command}")
+        os.system(f"{command}")
+        log.success(statement)
+
+    except FileNotFoundError as error:
+        log.info("\n-------------------------------------")
+        log.error(f"Something has gone wrong in executing {command}")
+        log.error(f" - {error}")
+        log.info("ARM UI currently stopped, fix error then restart ARM UI")
+        log.info("-------------------------------------")
+        exit
